requirements.txt
fastapi
uvicorn[standard]
sqlmodel
jinja2
python-multipart
python-jose[cryptography]
passlib[argon2]
slowapi
psutil
qrcode[pil]
pydantic-settings
python-dotenv